Implemented Socket Rooms First Stage
The player still plays against three computers, but now the Creation Stage and Adding Players sections are separated from the
game starting and the players are added via endpoints on the server.  Updates are also broadcast to a room, and not just back to the player
so multiple players can recieve information about a room.  First stage toward getting multiplayer working is now done